Move corsOptions to cofig file, delete views/subdir, delete  new-page.html
