+++ routes/root.js
@@ -6,12 +6,4 @@ router.get('^/$|/index(.html)?', (req, res) => {
   res.sendFile(path.join(__dirname, '..', 'views', 'index.html'))
 })
 
-router.get('/new-page(.html)?', (req, res) => {
-  res.sendFile(path.join(__dirname, '..', 'views', 'new-page.html'))
-})
-
-router.get('/old-page(.html)?', (req, res) => {
-  res.redirect(301, '/new-page.html') //302 by default
-})
-
 module.exports = router
