+++ server.js
@@ -2,6 +2,7 @@ const express = require('express')
 const app = express()
 const path = require('path')
 const cors = require('cors')
+const corsOptions = require('./config/corsOptions')
 const { logger } = require('./middleware/logEvents')
 const errorHandler = require('./middleware/errorHandler')
 const PORT = process.env.PORT || 3500 
@@ -10,17 +11,6 @@ const PORT = process.env.PORT || 3500
 app.use(logger)
 
 //Cross Origin Resourse Sharing
-const whitelist = ['https://www.yoursite.com', 'http://127.0.0.1:5500', 'http://localhost:3500']
-const corsOptions = {
-  origin: (origin, callback) => {
-    if(whitelist.indexOf(origin !== -1 || !origin)){
-      callback(null, true)
-    }else{
-      callback(new Error('Not allowed by CORS'))
-    }
-  },
-  optionsSuccessStatus: 200
-}
 app.use(cors(corsOptions))
 
 //Middleware to handle form data
@@ -31,10 +21,8 @@ app.use(express.json())
 
 //Serve static files
 app.use('/', express.static(path.join(__dirname, '/public')))
-app.use('/subdir', express.static(path.join(__dirname, '/public')))
 
 app.use('/', require('./routes/root'))
-app.use('/subdir', require('./routes/subdir'))
 app.use('/employees', require('./routes/api/employees'))
 
 //Route Handlers
