--- views/new-page.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-  <head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>New Page</title>
-    <link rel="stylesheet" href="css/style.css" />
-  </head>
-
-  <body>
-    <h1>New Page</h1>
-    <br>
-    <img src="img/img1.jpg">
-  </body>
-
-</html>
