--- views/subdir/test.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-  <head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Test Page</title>
-  </head>
-
-  <body>
-    <h1>Test Page in /subdir/</h1>
-  </body>
-
-</html>
